Rework of everything
Made completely new gravity system, because I didn't really enjoy the other one, so I decided to start all over again. In the end I think it turned out better than the last one but there is a long way for it to be perfect, which will be made later in development. I think that is the final result of these tasks.